update to 1.20

+++ plugins/finale-paper/src/main/java/com/github/maxopoly/finale/combat/knockback/KnockbackStrategy.java
@@ -8,6 +8,6 @@ import org.bukkit.util.Vector;
 
 public interface KnockbackStrategy {
 
-	void handleKnockback(Player attacker, Entity victim, int knockbackLevel);
+	void handleKnockback(Player attacker, Entity victim, float knockbackLevel);
 
 }
